fix: correctness wave across sort, IPC, C data, joins, arrays and dtypes
- `sort` now honours `stable`, so a multi-key sort stops returning wrong
  orders; stable sorts route to radix, which is stable for free.
- IPC: a sliced column no longer round-trips as the parent's first rows,
  compressed bodies are rejected, and delta dictionaries append.
- C data: resolve the "null count not computed" sentinel on import, stop the
  array stream double-freeing and leaking its dictionary, and check
  `n_buffers` before indexing the buffer array.
- Outer, semi and anti joins were wrong across probe morsels, and MARK declared
  columns it did not carry.
- Arrays: a slice reported the parent's null count, equality compared bitmaps
  rather than null positions, and device transfer dropped a sliced offset.
- `is_primitive()` said bool and `byte_width()` aborted on it; sliced kernel
  inputs had their nulls shifted by the offset.
- Split the expression layer into two lanes, and give string predicates a
  scalar-pattern entry point so a constant operand is not splatted per row.
- Gate binary size in CI and fix the documentation site build.

+++ marrow/tests/test_buffers_gpu.mojo
@@ -47,8 +47,8 @@ def test_buffer_host_builder() raises:
     """Buffer.alloc_host + to_immutable produce a valid HOST buffer."""
     var ctx = DeviceContext()
     var b = Buffer.alloc_host[DType.uint8](ctx, 64)
-    b.unsafe_set(0, 7)
-    b.unsafe_set(1, 13)
+    b.unsafe_set(0, UInt8(7))
+    b.unsafe_set(1, UInt8(13))
     var buf = b.to_immutable()
     assert_true(buf.is_host())
     assert_true(buf.is_cpu())
@@ -68,8 +68,8 @@ def test_buffer_to_cpu_round_trip() raises:
     """Upload a CPU buffer to GPU then download back; data is preserved."""
     var ctx = DeviceContext()
     var builder = Buffer.alloc_zeroed[DType.uint8](64)
-    builder.unsafe_set(0, 42)
-    builder.unsafe_set(1, 99)
+    builder.unsafe_set(0, UInt8(42))
+    builder.unsafe_set(1, UInt8(99))
     var cpu_buf = builder.to_immutable()
 
     var dev_buf = cpu_buf.to_device(ctx)
